restructure settings.py; deal with the import issue of examples
