--- dmff/admp/__init__.py
@@ -1,1 +1,0 @@
-import dmff.admp.settings
