+++ dmff/admp/recip.py
@@ -1,10 +1,9 @@
-#!/usr/bin/env python
-import sys
+
 import numpy as np
 import jax.numpy as jnp
 import jax.scipy as jsp
-from jax import jit, grad, value_and_grad
-from dmff.admp.settings import DO_JIT, jit_condition
+from jax import jit
+from dmff.settings import DO_JIT
 from dmff.admp.pme import DIELECTRIC
 
 # for debug
@@ -461,91 +460,91 @@ def Ck_10(ksq, kappa, V):
     return sqrt_pi*jnp.pi/2/V*kappa**7 * f / 1260
 
 
-def validation(pdb):
-    jnp.set_printoptions(precision=32, suppress=True)
-    xml = 'mpidwater.xml'
-    pdbinfo = read_pdb(pdb)
-    serials = pdbinfo['serials']
-    names = pdbinfo['names']
-    resNames = pdbinfo['resNames']
-    resSeqs = pdbinfo['resSeqs']
-    positions = pdbinfo['positions']
-    box = pdbinfo['box'] # a, b, c, α, β, γ
-    charges = pdbinfo['charges']
-    positions = jnp.asarray(positions)
-    lx, ly, lz, _, _, _ = box
-    box = jnp.eye(3)*jnp.array([lx, ly, lz])
+# def validation(pdb):
+#     jnp.set_printoptions(precision=32, suppress=True)
+#     xml = 'mpidwater.xml'
+#     pdbinfo = read_pdb(pdb)
+#     serials = pdbinfo['serials']
+#     names = pdbinfo['names']
+#     resNames = pdbinfo['resNames']
+#     resSeqs = pdbinfo['resSeqs']
+#     positions = pdbinfo['positions']
+#     box = pdbinfo['box'] # a, b, c, α, β, γ
+#     charges = pdbinfo['charges']
+#     positions = jnp.asarray(positions)
+#     lx, ly, lz, _, _, _ = box
+#     box = jnp.eye(3)*jnp.array([lx, ly, lz])
 
-    rc = 4  # in Angstrom
-    ethresh = 1e-4
+#     rc = 4  # in Angstrom
+#     ethresh = 1e-4
 
-    n_atoms = len(serials)
+#     n_atoms = len(serials)
 
-    atomTemplate, residueTemplate = read_xml(xml)
-    atomDicts, residueDicts = init_residues(serials, names, resNames, resSeqs, positions, charges, atomTemplate, residueTemplate)
+#     atomTemplate, residueTemplate = read_xml(xml)
+#     atomDicts, residueDicts = init_residues(serials, names, resNames, resSeqs, positions, charges, atomTemplate, residueTemplate)
 
-    Q = np.vstack(
-        [(atom.c0, atom.dX*10, atom.dY*10, atom.dZ*10, atom.qXX*300, atom.qYY*300, atom.qZZ*300, atom.qXY*300, atom.qXZ*300, atom.qYZ*300) for atom in atomDicts.values()]
-    )
-    Q = jnp.array(Q)
-    Q_local = convert_cart2harm(Q, 2)
-    axis_type = np.array(
-        [atom.axisType for atom in atomDicts.values()]
-    )
-    axis_indices = np.vstack(
-        [atom.axis_indices for atom in atomDicts.values()]
-    )
-    covalent_map = assemble_covalent(residueDicts, n_atoms)
-
-
-    # invoke pme energy calculator
-    # energy_pme(positions, box, Q_local, axis_type, axis_indices, nbr.idx, 2)
-    lmax = 2
-    kappa, K1, K2, K3 = setup_ewald_parameters(rc, ethresh, box)
-    # for debugging
-    kappa = 0.657065221219616
-    construct_local_frames_fn = generate_construct_local_frames(axis_type, axis_indices)
-    local_frames = construct_local_frames_fn(positions, box)
-    Q_global = rot_local2global(Q_local, local_frames, lmax)
-
-    pme_order = 6
-    energy_force_pme_recip = value_and_grad(generate_pme_recip(Ck_1, kappa, False, pme_order, K1, K2, K3, lmax))
-    energy_force_pme_recip(positions, box, Q_global)
-    print('ok')
-    E, F = energy_force_pme_recip(positions, box, Q_global)
-    print(E)
-
-    # construct the C list
-    c_list = np.zeros((3,n_atoms))
-    nmol=int(n_atoms/3)
-    for i in range(nmol):
-        a = i*3
-        b = i*3+1
-        c = i*3+2
-        c_list[0][a]=37.19677405
-        c_list[0][b]=7.6111103
-        c_list[0][c]=7.6111103
-        c_list[1][a]=85.26810658
-        c_list[1][b]=11.90220148
-        c_list[1][c]=11.90220148
-        c_list[2][a]=134.44874488
-        c_list[2][b]=15.05074749
-        c_list[2][c]=15.05074749
-    energy_force_d6_recip = value_and_grad(generate_pme_recip(Ck_6, kappa, True, pme_order, K1, K2, K3, 0))
-    energy_force_d8_recip = value_and_grad(generate_pme_recip(Ck_8, kappa, True, pme_order, K1, K2, K3, 0))
-    energy_force_d10_recip = value_and_grad(generate_pme_recip(Ck_10, kappa, True, pme_order, K1, K2, K3, 0))
-    E6, F6 = energy_force_d6_recip(positions, box, c_list[0, :, jnp.newaxis])
-    E8, F6 = energy_force_d8_recip(positions, box, c_list[1, :, jnp.newaxis])
-    E10, F10 = energy_force_d10_recip(positions, box, c_list[2, :, jnp.newaxis])
-    print('ok')
-    E6, F6 = energy_force_d6_recip(positions, box, c_list[0, :, jnp.newaxis])
-    E8, F6 = energy_force_d8_recip(positions, box, c_list[1, :, jnp.newaxis])
-    E10, F10 = energy_force_d10_recip(positions, box, c_list[2, :, jnp.newaxis])
-    print(E6, E8, E10)
-    print(E6 + E8 + E10)
-    return
+#     Q = np.vstack(
+#         [(atom.c0, atom.dX*10, atom.dY*10, atom.dZ*10, atom.qXX*300, atom.qYY*300, atom.qZZ*300, atom.qXY*300, atom.qXZ*300, atom.qYZ*300) for atom in atomDicts.values()]
+#     )
+#     Q = jnp.array(Q)
+#     Q_local = convert_cart2harm(Q, 2)
+#     axis_type = np.array(
+#         [atom.axisType for atom in atomDicts.values()]
+#     )
+#     axis_indices = np.vstack(
+#         [atom.axis_indices for atom in atomDicts.values()]
+#     )
+#     covalent_map = assemble_covalent(residueDicts, n_atoms)
 
 
-# validation code
-if __name__ == '__main__':
-    validation(sys.argv[1])
+#     # invoke pme energy calculator
+#     # energy_pme(positions, box, Q_local, axis_type, axis_indices, nbr.idx, 2)
+#     lmax = 2
+#     kappa, K1, K2, K3 = setup_ewald_parameters(rc, ethresh, box)
+#     # for debugging
+#     kappa = 0.657065221219616
+#     construct_local_frames_fn = generate_construct_local_frames(axis_type, axis_indices)
+#     local_frames = construct_local_frames_fn(positions, box)
+#     Q_global = rot_local2global(Q_local, local_frames, lmax)
+
+#     pme_order = 6
+#     energy_force_pme_recip = value_and_grad(generate_pme_recip(Ck_1, kappa, False, pme_order, K1, K2, K3, lmax))
+#     energy_force_pme_recip(positions, box, Q_global)
+#     print('ok')
+#     E, F = energy_force_pme_recip(positions, box, Q_global)
+#     print(E)
+
+#     # construct the C list
+#     c_list = np.zeros((3,n_atoms))
+#     nmol=int(n_atoms/3)
+#     for i in range(nmol):
+#         a = i*3
+#         b = i*3+1
+#         c = i*3+2
+#         c_list[0][a]=37.19677405
+#         c_list[0][b]=7.6111103
+#         c_list[0][c]=7.6111103
+#         c_list[1][a]=85.26810658
+#         c_list[1][b]=11.90220148
+#         c_list[1][c]=11.90220148
+#         c_list[2][a]=134.44874488
+#         c_list[2][b]=15.05074749
+#         c_list[2][c]=15.05074749
+#     energy_force_d6_recip = value_and_grad(generate_pme_recip(Ck_6, kappa, True, pme_order, K1, K2, K3, 0))
+#     energy_force_d8_recip = value_and_grad(generate_pme_recip(Ck_8, kappa, True, pme_order, K1, K2, K3, 0))
+#     energy_force_d10_recip = value_and_grad(generate_pme_recip(Ck_10, kappa, True, pme_order, K1, K2, K3, 0))
+#     E6, F6 = energy_force_d6_recip(positions, box, c_list[0, :, jnp.newaxis])
+#     E8, F6 = energy_force_d8_recip(positions, box, c_list[1, :, jnp.newaxis])
+#     E10, F10 = energy_force_d10_recip(positions, box, c_list[2, :, jnp.newaxis])
+#     print('ok')
+#     E6, F6 = energy_force_d6_recip(positions, box, c_list[0, :, jnp.newaxis])
+#     E8, F6 = energy_force_d8_recip(positions, box, c_list[1, :, jnp.newaxis])
+#     E10, F10 = energy_force_d10_recip(positions, box, c_list[2, :, jnp.newaxis])
+#     print(E6, E8, E10)
+#     print(E6 + E8 + E10)
+#     return
+
+
+# # validation code
+# if __name__ == '__main__':
+#     validation(sys.argv[1])
