+++ dmff/admp/spatial.py
@@ -1,9 +1,10 @@
-#!/usr/bin/env python
+
 import jax.numpy as jnp
 from jax import vmap, jit
 import numpy as np
 from functools import partial
-from dmff.admp.settings import DO_JIT, jit_condition
+from dmff.settings import DO_JIT
+from dmff.utils import jit_condition
 
 # This module deals with spatial geometric operations, mainly including:
 # 1. PBC related operations
