+++ dmff/utils.py
@@ -1,0 +1,10 @@
+from dmff.settings import DO_JIT
+from jax import jit
+
+def jit_condition(*args, **kwargs):
+    def jit_deco(func):
+        if DO_JIT:
+            return jit(func, *args, **kwargs)
+        else:
+            return func
+    return jit_deco
