--- examples/water_pol_1024/admp
@@ -1,1 +1,0 @@
-../../admp/
